Initial Commit
This is lab06 working fully, but there appear to be some bugs. Sometimes
the command will successfully be run but the output won't match.
Sometimes commands entered are not run successfully.  I can't get any
breakpoints to be successfully hit in the tasks UART TX and UART RX. The
breakpoints work fine if you put them in other places like the uartdrv.c
driver functions.